requirements.txt
opencv-python>=4.5.0
numpy>=1.21.0
tensorflow>=2.8.0
torch>=1.10.0
torchvision>=0.11.0
av>=9.0.0
aiohttp>=3.8.0
aiohttp-jinja2>=1.5.0
jinja2>=3.0.0
websockets>=10.0
redis>=4.0.0
pyyaml>=6.0
prometheus_client>=0.13.0
psutil>=5.9.0
python-nginx==1.5.7
m3u8>=3.0.0
python-dotenv==1.0.1  # For environment variables
